requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
jinja2>=3.1
python-multipart>=0.0.9
python-dotenv>=1.0
faster-whisper>=1.0.3
pillow>=10.2
opencv-python>=4.9
numpy>=1.26
httpx>=0.27
openai>=1.30
